+++ core/src/com/sleepware/GameObjects/FallingScore.java
@@ -24,8 +24,8 @@ public class FallingScore extends RotatingScrollable {
 	}
 	
 	@Override
-	public void setLevelAttributes(int scrollSpeed) {
-		baseYVelocity=-scrollSpeed;
+	public void setLevelAttributes(float scrollSpeed) {
+		baseYVelocity=(int)-scrollSpeed;
 	}
 	
 	@Override
